% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{merge_overlaps}
\alias{merge_overlaps}
\title{Merge overlapping / adjacent intervals.}
\usage{
merge_overlaps(inst_data)
}
\arguments{
\item{inst_data}{data.frame with `patient_id`, `start`, `end`.}
}
\value{
Data.frame with merged intervals per patient.
}
\description{
Merge overlapping / adjacent intervals.
}
