% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{count_institutional_days}
\alias{count_institutional_days}
\title{Count institutional days inside the window.}
\usage{
count_institutional_days(inst_data, win_data)
}
\arguments{
\item{inst_data}{data.frame with merged intervals (`patient_id`, `start`, `end`).}

\item{win_data}{data.frame from `apply_window_and_death`
(must contain `patient_id`, `window_start`, `window_end`).}
}
\value{
Data.frame with `patient_id`, `institutional_days`.
}
\description{
Count institutional days inside the window.
}
