% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{plot_dah_distribution}
\alias{plot_dah_distribution}
\title{Plot the DAH distribution.}
\usage{
plot_dah_distribution(data, window_days = 30L)
}
\arguments{
\item{data}{per‑patient result from `compute_dah`.}

\item{window_days}{observation window length (default = 30).}
}
\value{
A `ggplot2` object.
}
\description{
Plot the DAH distribution.
}
